Fully implemented and tested NestedDictionary class
-- Wrote tests for all valid use cases for all class methods and used them to verify that class works as expected.

As part of the verification, a niche case of unexpected behavior was identified and fixed:
The class now treats non-top-level empty sub-dictionaries as valid terminal values. That is, paths to empty sub-dictionaries will now be returned by terminal paths scanning method and, consequently, handled by all methods that use that method.