Removed DataConverters
DataConverters were designed for libraries that configure runtimes through .yaml configs with the potential need to support multiple input types. Throughout development, Ataraxis project instead went with strict data-typing and heavily relying on numpy, numba and polars. Therefore, this functionality is not really needed for any of our project and can be safely factored out of the library. DataConverters will be refactored into a different / standalone library.

The primary reason for this change is to keep the core library on which many other upstream libraries depend as light as possible.